Change to score map function name, move score function to new utils file, generalize single_animation function
